+++ docs/src/pages/components/selects/BasicSelect.js
@@ -5,17 +5,13 @@ import MenuItem from '@material-ui/core/MenuItem';
 import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 
-const useStyles = makeStyles((theme) => ({
-  formControl: {
-    margin: theme.spacing(1),
+const useStyles = makeStyles({
+  root: {
     minWidth: 120,
   },
-  selectEmpty: {
-    marginTop: theme.spacing(2),
-  },
-}));
+});
 
-export default function SimpleSelect() {
+export default function BasicSelect() {
   const classes = useStyles();
   const [age, setAge] = React.useState('');
 
@@ -24,19 +20,15 @@ export default function SimpleSelect() {
   };
 
   return (
-    <div>
-      <FormControl required className={classes.formControl}>
-        <InputLabel id="demo-simple-select-required-label">Age</InputLabel>
+    <div className={classes.root}>
+      <FormControl fullWidth>
+        <InputLabel id="demo-simple-select-label">Age</InputLabel>
         <Select
-          labelId="demo-simple-select-required-label"
-          id="demo-simple-select-required"
+          labelId="demo-simple-select-label"
+          id="demo-simple-select"
           value={age}
           onChange={handleChange}
-          className={classes.selectEmpty}
         >
-          <MenuItem value="">
-            <em>None</em>
-          </MenuItem>
           <MenuItem value={10}>Ten</MenuItem>
           <MenuItem value={20}>Twenty</MenuItem>
           <MenuItem value={30}>Thirty</MenuItem>
