+++ docs/src/pages/components/selects/MultipleSelectNative.js
@@ -25,7 +25,7 @@ const names = [
   'Kelly Snyder',
 ];
 
-export default function MultipleSelect() {
+export default function MultipleSelectNative() {
   const classes = useStyles();
   const [personName, setPersonName] = React.useState([]);
   const handleChangeMultiple = (event) => {
