+++ docs/src/pages/components/selects/NativeSelect.tsx
@@ -4,19 +4,18 @@ import InputLabel from '@material-ui/core/InputLabel';
 import FormControl from '@material-ui/core/FormControl';
 import NativeSelect from '@material-ui/core/NativeSelect';
 
-const useStyles = makeStyles((theme) => ({
-  formControl: {
-    margin: theme.spacing(1),
+const useStyles = makeStyles({
+  root: {
     minWidth: 120,
   },
-}));
+});
 
-export default function NativeSelects() {
+export default function NativeSelectDemo() {
   const classes = useStyles();
 
   return (
-    <div>
-      <FormControl className={classes.formControl}>
+    <div className={classes.root}>
+      <FormControl fullWidth>
         <InputLabel htmlFor="uncontrolled-native">Age</InputLabel>
         <NativeSelect
           defaultValue={30}
