+++ docs/src/pages/components/selects/SelectAutoWidth.tsx
@@ -17,7 +17,7 @@ const useStyles = makeStyles((theme: Theme) =>
   }),
 );
 
-export default function SimpleSelect() {
+export default function SelectAutoWidth() {
   const classes = useStyles();
   const [age, setAge] = React.useState('');
