+++ docs/src/pages/components/selects/SelectOtherProps.js
@@ -16,7 +16,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function SimpleSelect() {
+export default function SelectOtherProps() {
   const classes = useStyles();
   const [age, setAge] = React.useState('');
